task 104 commit 8

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int i, j;
+	long i, j;
 	int f = 0;
 	long h = 0;
 	long k1 = 0;
